+++ connectomix/tests/seeds/example_seeds_for_roiToRoi.tsv
@@ -1,0 +1,53 @@
+Left Dentate Nuclei	-15	-61	-36
+Right Dentate Nuclei	15	-61	-36
+Left Crus I	-36	-74	-46
+Right Crus I	42	-72	-48
+Left Crus II	-34	-66	-32
+Right Crus II	28	-68	-30
+Left Lobule VIIa	-6	-86	-32
+Right Lobule VIIa	6	-82	-28
+Left Lobule V	-14	-48	-14
+Right Lobule V	14	-46	-14
+Left Lobule VI	-20	-60	-16
+Right Lobule VI	18	-58	-14
+Left Lobule VIII	-22	-52	-60
+Right Lobule VIII	24	-62	-58
+L Aud	-54	-9	32
+R Aud	56	-5	31
+Striate	0	-76	11
+L DMN	-47	-65	31
+Med DMN	0	-56	29
+Front DMN	0	53	7
+R DMN	54	-59	28
+Occ post	2	-93	-1
+Motor	0	-22	61
+R DLPFC	44	21	42
+R Front pol	39	56	1
+R Par	50	-53	46
+R Post Temp	63	-33	-8
+Basal	-17	6	5
+L Par	-42	-61	47
+L DLPFC	-45	18	40
+L Front pol	-42	51	-1
+L IPS	-25	-68	43
+R IPS	33	-61	44
+L LOC	-32	-87	9
+Vis	-26	-72	-11
+R LOC	37	-80	7
+D ACC	-28	47	22
+V ACC	0	31	24
+R A Ins	28	48	25
+L STS	-56	-21	14
+R STS	57	-15	13
+L TPJ	-55	-35	-2
+Broca	-50	26	3
+Sup Front S	-4	15	64
+R TPJ	52	-31	-2
+R Pars Op	53	30	1
+Cereb	4	-61	-18
+Dors PCC	-1	-60	58
+L Ins	-38	15	2
+Cing	1	8	48
+R Ins	39	16	2
+L Ant IPS	-49	-31	42
+R Ant IPS	50	-27	43
