+++ connectomix/tests/seeds/example_seeds_for_roiToVoxel.tsv
@@ -1,0 +1,2 @@
+Med DMN	0	-56	29
+Front DMN	0	53	7
