+++ connectomix/version.py
@@ -1,0 +1,2 @@
+# Define the version number
+__version__ = "dev"
