+++ requirements.txt
@@ -1,9 +1,51 @@
-bids
-matplotlib
-nibabel
-nilearn
-numpy
-pandas
-pathlib
-scipy
-statsmodels
+astor==0.8.1
+attrs==24.3.0
+bids-validator==1.14.7.post0
+bidsschematools==1.0.0
+certifi==2024.12.14
+charset-normalizer==3.4.1
+click==8.1.8
+contourpy==1.3.1
+cycler==0.12.1
+docopt==0.6.2
+fonttools==4.55.3
+formulaic==0.5.2
+fsspec==2024.12.0
+greenlet==3.1.1
+idna==3.10
+importlib_resources==6.4.5
+interface-meta==1.3.0
+joblib==1.4.2
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+kiwisolver==1.4.8
+lxml==5.3.0
+matplotlib==3.10.0
+nibabel==5.3.2
+nilearn==0.11.1
+num2words==0.5.14
+numpy==2.2.1
+packaging==24.2
+pandas==2.2.3
+pathlib==1.0.1
+patsy==1.0.1
+pillow==11.0.0
+pybids==0.18.1
+pyparsing==3.2.0
+python-dateutil==2.9.0.post0
+pytz==2024.2
+PyYAML==6.0.2
+referencing==0.35.1
+requests==2.32.3
+rpds-py==0.22.3
+scikit-learn==1.6.0
+scipy==1.14.1
+six==1.17.0
+SQLAlchemy==2.0.36
+statsmodels==0.14.4
+threadpoolctl==3.5.0
+typing_extensions==4.12.2
+tzdata==2024.2
+universal_pathlib==0.2.6
+urllib3==2.3.0
+wrapt==1.17.0
